reporting: re-organise sources into seperate modules
