+++ compiler/hash-ast-passes/src/diagnostics/warning.rs
@@ -1,7 +1,8 @@
 //! Hash AST semantic analysis warning diagnostic definitions.
 
-use hash_reporting::reporting::{
-    Report, ReportBuilder, ReportCodeBlock, ReportElement, ReportKind, ReportNote, ReportNoteKind,
+use hash_reporting::{
+    builder::ReportBuilder,
+    report::{Report, ReportCodeBlock, ReportElement, ReportKind, ReportNote, ReportNoteKind},
 };
 use hash_source::location::SourceLocation;
