+++ compiler/hash-reporting/src/lib.rs
@@ -1,4 +1,6 @@
 //! Hash Compiler error and warning reporting module
+pub mod builder;
 pub mod errors;
 pub mod highlight;
-pub mod reporting;
+pub mod report;
+pub mod writer;
